backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
langchain==0.1.6
langchain-openai==0.0.5
langgraph==0.0.28
openai==1.12.0
httpx==0.26.0